MDEV-22166 MIGRATE PARTITION: move out partition into a table
New ALTER_PARTITION_EXTRACT command for fast_alter_partition_table()
is done in alter_partition_extract() function which basically does
ha_rename_table().

Partition to extract is marked with the same flag as dropped
partition: PART_TO_BE_DROPPED. Note that we cannot have multiple
partitioning commands in one ALTER.

For DDL logging basically the principle is the same as for other
fast_alter_partition_table() commands. The only difference is that it
integrates late Atomic DDL functions and introduces additional phase
of WFRM_BACKUP_ORIGINAL. That is required for binlog consistency
because otherwise we could not revert back after WFRM_INSTALL_SHADOW
is done. And before DDL log is complete if we crash or fail the
altered table will be already new but binlog will miss that ALTER
command. There is a patch that helps to binlog command in
DDL_LOG_RENAME_ACTION but it turns out that WFRM_BACKUP_ORIGINAL is
simpler and more robust solution.

Test cases added to:

  parts.alter_table \
  versioning.partition \
  atomic.alter_partition