MDEV-26364 Add Spider server variable for disabling GROUP BY handler
Add a server variable, to the Spider storage engine, that disables
the use of the GROUP BY handler.

This provides users with a unified workaround for the bugs like
MDEV-26013. It is also useful for testing purposes because one can
specify the execution path that he want to test.

In the Spider, an execution is called "direct" if it uses the GROUP
BY handler (e.g., direct aggregation, direct join, ...). So, we name
the variable to be spider_enable_direct_select.